+++ src/qml/qml/qqmltypeloader.cpp
@@ -1062,7 +1062,7 @@ void QQmlTypeLoader::Blob::dependencyComplete(const QQmlDataBlob::Ptr &blob)
 
 bool QQmlTypeLoader::Blob::loadDependentImports(
         const QList<QQmlDirParser::Import> &imports, const QString &qualifier,
-        QTypeRevision version, quint16 precedence, QQmlImports::ImportFlags flags,
+        QTypeRevision version, quint8 precedence, QQmlImports::ImportFlags flags,
         QList<QQmlError> *errors)
 {
     assertTypeLoaderThread();
